+++ .gitignore
@@ -59,3 +59,4 @@ typings/
 
 public/bundle.js 
 package-lock.json
+app/components/Participants.js.orig
