+++ public/css/bootstrap.css
@@ -2518,7 +2518,7 @@ input[type="search"] {
 }
 input[type="radio"],
 input[type="checkbox"] {
-  margin: 4px 0 0;
+  margin: 4px 4px 0;
   margin-top: 1px \9;
   line-height: normal;
 }
